[OPT]update to zy_compass_4.0:     1. support quantization aware training     2. add operators: cacheupdate, dummy, If, Loop     3. enhance global calibration: svd_based_quant     4. fix bugs

+++ AIPUBuilder/Optimizer/ops/activation.py
@@ -1,5 +1,5 @@
 # SPDX-License-Identifier: Apache-2.0
-# Copyright © 2022-2024 Arm Technology (China) Co. Ltd.
+# Copyright © 2022-2025 Arm Technology (China) Co. Ltd.
 
 from AIPUBuilder.Optimizer.utils import *
 from AIPUBuilder.Optimizer.framework import *
@@ -60,7 +60,7 @@ def unknown_activation(self, *args):
     else:
         func = self.get_attrs('lambda_func', optional=True, default_value=None)
         if func is not None:
-            out.betensor = func(inp.betensor)
+            out.betensor = func(inp.betensor.float())
         else:
             OPT_WARN(f"Activation op method=UNKNOWN does not support float forward, and the output.tensor will directly use input.tensor")
             out.betensor = inp.betensor
